Ran successfully on AGB2015 rasters
